boot: uboot-armsr: update to U-Boot 2025.04
This U-Boot version is used to generate a 'reference'
U-Boot binary for QEMU's 'virt' machine on both
armv7 and armv8.

It has not been updated since EFI support was merged
into the then-armvirt target, so we should bring it up
to the latest version.

The 'mkeficapsule' tool is disabled due to a host-side
dependency on GnuTLS.

Signed-off-by: Mathew McBride <matt@traverse.com.au>
Link: https://github.com/openwrt/openwrt/pull/18862
Signed-off-by: Robert Marko <robimarko@gmail.com>